Reset after client disconnects, HEX bugfix
We now reset the module on disconnect, this is to prevent a flood
of NFail messages if a notification had been turned on.
A better options may be to just not update anything when not
connected to a client, but the behavior would have to be tested
when accessing with multiple clients.
Also bug fixed the HEX to int code.